# spec/openapi.yaml
openapi: 3.0.0
servers:
# Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/ProtocolOne/Authentication/1.0.0
  - url: 'https://auth-one.protocol.one'
info:
  description: The Authentication API enables you to manage all aspects of user identity when you use Auth-One. It offers endpoints so your users can log in, sign up, log out, access APIs, and more.
  version: 1.0.0
  title: ProtocolOne Authentication API
  termsOfService: ''
  contact:
    email: vadim.sabirov@protocol.one
    name: Vadim Sabirov
    url: 'https://protocol.one'
...
paths:
...
  /userinfo:
    get:
      tags:
      - Get User Info
      summary: 'Get user info by Access Token'
      description: 'Given the Access Token obtained during login, this endpoint returns a user''s profile.'
      parameters:
        - name: access_token
          in: query
          required: true
          schema:
            type: string
          description: 'The `Access Token` obtained during login'
      operationId: userInfo
      responses:
        '200':
          description: 'Return user info.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
  /token/refresh:
    get:
      tags:
      - JWT Token
      summary: 'Get new user Access Token by Refresh Token'
      description: ''
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: 'Refresh token'
      operationId: tokenRefresh
      responses:
        '200':
          description: 'User Access Tokens.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - BearerAuth: []
  /token/ott:
    get:
      tags:
      - JWT Token
      summary: 'Get user Access Token by One-Time Token'
      description: ''
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: 'One-Time Token (for example, the parameter `auth_one_ott` passed during authentication with redirect)'
      operationId: tokenOTT
      responses:
        '200':
          description: 'User Access Tokens'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
...
  /passwordless/start:
    post:
      tags:
      - Passwordless
      summary: 'Send a verification code using email or sms'
      description:
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: >-
            How to send the code to the user. Use `email` to send the
            code using email, or `sms` to use SMS
      operationId: passwordlessStart
      responses:
        '200':
          description: 'Return One-Time Token for verification method'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OneTimeToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /passwordless/verify:
    post:
      tags:
      - Passwordless
      summary: 'Verify passwordless authentication code'
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: 'The `client_id` of your application'
        - name: connection
          in: query
          required: true
          schema:
            type: string
          description: >-
            How to send the code to the user. Use `email` to send the
            code using email, or `sms` to use SMS
        - name: verification_code
          in: query
          required: true
          schema:
            type: string
          description: 'Verification code from sms or email (One-Time Code)'
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: 'One-Time Token'
      operationId: passwordlessVerify
      responses:
        '200':
          description: 'User Access Token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWTToken'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: bad_request
          description: 'Error code (for example: `mfa_request` for the MFA scenario or `password` for form field)'
        error_message:
          type: string
          example: The connection was not found
          description: 'Text of error or value for specific property (e.g. MFA one-time token value)'
    User:
      type: object
      properties:
        id:
          type: string
          example: b62klDf0HeiJdNMv8K263nfE
        email:
          type: string
          example: test@test.com
        email_verified:
          type: boolean
          example: false
        created_at:
          type: string
          example: "2016-11-02T15:05:21.311Z"
        updated_at:
          type: string
          example: "2016-12-05T15:15:40.545Z"
    JWTToken:
      type: object
      properties:
        access_token:
          type: string
          example: oDnPKc1a...vWE36Lh
          description: JWT token
        expires_in:
          type: integer
          format: int32
          example: 3600
          description: UNIX timestamp
        id_token:
          type: string
          description: Refresh token
          example: oDnPKc1a...vWE36Lh
...
    OneTimeToken:
      type: object
      description: 'One-Time Token'
      properties:
        token:
          type: string
          example: oDnPKc1a...vWE36Lh
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: X-AUTH-ONE-TOKEN-{CLIENT_ID}
    BearerAuth:
      type: http
      scheme: Bearer
      bearerFormat: JWT
...
security:
  - CookieAuth: []
  - BearerAuth: []